services/agents_service/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agents_service"
version = "0.1.0"
description = "Bibliotalk agents service"
requires-python = ">=3.11"
dependencies = [
  "bt_common>=0.1.0",
  "bt_store>=0.1.0",
  "fastapi>=0.115.0",
  "google-adk>=1.25.0",
  "google-genai>=1.0.0",
  "httpx[socks]>=0.27.0",
  "pydantic>=2.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.6.9",
]

[tool.setuptools]
packages = [
  "agents_service",
  "agents_service.agent",
  "agents_service.agent.providers",
  "agents_service.agent.tools",
  "agents_service.api",
  "agents_service.audit",
  "agents_service.live",
  "agents_service.models",
]

[tool.setuptools.package-dir]
"agents_service" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]
ignore = ["E501"]

[tool.uv.sources]
bt-common = { workspace = true }
bt-store = { workspace = true }
